+++ libs/mqtt/src/contracts/index.ts
@@ -1,0 +1,2 @@
+export * from './mqtt';
+export const LIB_MQTT = Symbol('mqtt');
