+++ libs/mqtt/src/decorators/index.ts
@@ -1,0 +1,2 @@
+export * from './inject-mqtt.decorator';
+export * from './mqtt.decorator';
