+++ libs/utilities/src/contracts/index.ts
@@ -8,5 +8,4 @@ export * from './interfaces';
 export * from './lifecycle';
 export * from './logger';
 export * from './meta';
-export * from './mqtt';
 export * from './nx-workspace.dto';
