+++ libs/utilities/src/decorators/injectors/index.ts
@@ -1,4 +1,3 @@
 export * from './inject-cache.decorator';
 export * from './inject-config.decorator';
 export * from './inject-logger.decorator';
-export * from './inject-mqtt.decorator';
