+++ libs/utilities/src/modules/index.ts
@@ -1,2 +1,1 @@
-export * from './mqtt.module';
 export * from './utilities.module';
