+++ libs/utilities/src/services/index.ts
@@ -4,6 +4,5 @@ export * from './explorers';
 export * from './fetch';
 export * from './json-filter.service';
 export * from './lifecycle.service';
-export * from './mqtt';
 export * from './providers';
 export * from './workspace.service';
